fixed int value overflow issuss for DCU

+++ tensortype/dcu_kernels/add_bias.cpp
@@ -52,7 +52,7 @@ int kr_add_bias<__half>(const __half* in, const __half* bias, __half* out, int l
 // ===============================================
 template<typename T>
 __global__ void add_broadcast(const T* in, const T* bias, T* out, int length, int inter, int feature) {
-    int i = blockIdx.x * blockDim.x + threadIdx.x;
+    size_t i = blockIdx.x * blockDim.x + threadIdx.x;
     if ( i >= length * feature * inter) {
         return;
     }
@@ -66,8 +66,8 @@ __global__ void add_broadcast(const T* in, const T* bias, T* out, int length, in
 }
 
 template <>
-int kr_add_broadcast<float>(const float* in, const float* bias, float* out, int length, int inter, int feature, hipStream_t stream) {
-    int nElementNumber = length * feature * inter;
+int kr_add_broadcast<float>(const float* in, const float* bias, float* out, int length, size_t inter, size_t feature, hipStream_t stream) {
+    size_t nElementNumber = length * feature * inter;
 
     dim3 block_size(256);
 	dim3 num_of_blocks((nElementNumber + block_size.x - 1) / block_size.x);
@@ -83,8 +83,8 @@ int kr_add_broadcast<float>(const float* in, const float* bias, float* out, int 
 }
 
 template <>
-int kr_add_broadcast<__half>(const __half* in, const __half* bias, __half* out, int length, int inter, int feature, hipStream_t stream) {
-    int nElementNumber = length * feature * inter;
+int kr_add_broadcast<__half>(const __half* in, const __half* bias, __half* out, int length, size_t inter, size_t feature, hipStream_t stream) {
+    size_t nElementNumber = length * feature * inter;
 
     dim3 block_size(256);
 	dim3 num_of_blocks((nElementNumber + block_size.x - 1) / block_size.x);
